requirements.txt
streamlit
feedparser
beautifulsoup4
pandas
lxml
pillow
